Update Azure
#37 

> Following the assumption that interconnection between the CPS and cloud services is designed based on the gateway concept, a middleware must be considered as a coupler.

In the previous chapter it was said that "the embedded gateway is not derived from the middleware concept" (I've noted it in #43 too, so I guess it should be adjusted earlier?)

> Removed  "called telemetry"

Depending on the context this could be wrong as the azure docs specify various kinds of data including but not limited to telemetry. I think it's best to skip this part.